openai-chatbot-backend/app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
openai>=1.0.0
python-dotenv==1.0.0
pydantic==1.10.7
websockets==11.0.3